move dbsettings.* from libmyth to libmythui
