+++ mythtv/libs/libmythui/dbsettings.cpp
