+++ mythtv/libs/libmythui/libmythui.pro
@@ -50,6 +50,7 @@ HEADERS += mythvrr.h
 HEADERS += mythcolourspace.h
 HEADERS += devices/mythinputdevicehandler.h
 HEADERS += mythuiprocedural.h
+HEADERS += dbsettings.h
 HEADERS += guistartup.h
 HEADERS += langsettings.h
 HEADERS += mediamonitor.h
@@ -93,6 +94,7 @@ SOURCES += mythvrr.cpp
 SOURCES += mythcolourspace.cpp
 SOURCES += devices/mythinputdevicehandler.cpp
 SOURCES += mythuiprocedural.cpp
+SOURCES += dbsettings.cpp
 SOURCES += guistartup.cpp
 SOURCES += langsettings.cpp
 SOURCES += mediamonitor.cpp
